Refactor ANMN dest_path: * move more generic methods from ANMNFileClassifier into FileClassifier base class * rename ANMNFileClassifier to MooringFileClassifier and move to lib/python * update tests (no longer need ANMN test_dest_path)
